Add CommodityCustomRepository And TestClass
getCommodities 메서드 테스트만 일단 구현

+++ src/main/java/backend/jangbogoProject/entity/category/controller/CategoryController.java
@@ -18,7 +18,7 @@ public class CategoryController {
     private final CategoryService categoryService;
 
     @PostMapping("/create")
-    public Long createCategory(@RequestBody @Valid CategoryRequestDTO categoryRequestDTO){
+    public int createCategory(@RequestBody @Valid CategoryRequestDTO categoryRequestDTO){
         return categoryService.create(categoryRequestDTO);
     }
